Fix A TON of severe EventBus problems.
Before, the EventBus associated each IEvent with a set of handlers as well as _one_ object that would serve as the handlers' implicit argument. However, the EventBus did not account for the fact that handlers could originate from multiple classes, each requiring a seperate instance to be properly invoked.
Also fixed ConcurrentModificationExceptions that would occur when the EventBus executed a handler that would need to load a new class into the JVM. This, in turn, would invoke SubscribingClassLoader, which would invoke the hook that automatically subscribed things to the eventbus...thus breaking the fail-fast behavior of the iterator used to iterate over the subscribed methods.